backend/requirements.txt
google-adk==0.3.0
google-generativeai==0.8.5
#google-cloud-aiplatform>=1.105.0,<2.0.0
python-dotenv>=1.0.0
uvicorn>=0.23.2
fastapi>=0.100.0
grpcio>=1.56.0
protobuf>=4.23.0
